Scanner is now built from a list of rules
This refactor should provide the highest re-usability of the
existing code when writing a subclass.  Keyword changes get
included automatically in the scanner rules, the subclass
can re-order the scanner rules or inject new ones in the sequence
without needing to wrap the whole method, or wrap the whole
body of the method without needing to rewrite the eval code.