Change javascript variable-access to be more like Perl counterpart
In the perl compiler, the variable comes from calling get_value
on the Namespace object.  While the namespace itself can't be
passed to javascript, a comparable design is to call that
method on the first argument to the function.

If the method is not set, it acts just like perl and looks up
a named field of the first argument.