Document the scanner_rules function group, and fix bug in keyword_map
While writing up the documentation, I realized that there was a bug
in keyword_map where if a subclass was the first to call
->keyword_map, it would cache the subclass's keywords in the parent.

I changed the design of scanner_rules to add a 4th array element that
holds a hashref of the named variables that should be made available
to the code in the eval.  Now, scanner_rules can call keyword_map once
and then pass it on to _build_scan_token_method who will build it into
the eval generically, avoiding a second call to keyword_map.

Since it is only called once and not cached user-visible, I also no
longer needed Const::Fast.