Introduce Formula object for more convenient access to API
The formula object holds onto the original string of text, the
parse tree, and the sets of functions and variables.  It can also
calculate the sets on demand when its been constructed from a new
parse tree.  This also makes it a nice place for the 'simplify'
method so that it can keep track of the details involved around that.